+++ Assets/LDtkUnity/Editor/CustomEditor/Importer/LDtkEditorCommandUpdater.cs
@@ -1,7 +1,5 @@
 ﻿using System.Diagnostics;
 using System.IO;
-using System.Linq;
-using System.Text.RegularExpressions;
 using UnityEditor;
 using UnityEngine;