+++ Assets/LDtkUnity/Editor/CustomEditor/SceneDrawer/ComponentDrawers/LDtkSceneDrawerEntity.cs
@@ -1,70 +1,69 @@
 ﻿using System;
-using UnityEditor;
+using System.Linq;
 using UnityEngine;
 
 namespace LDtkUnity.Editor
 {
+    /// <summary>
+    /// Draws entity stuff, no field stuff
+    /// </summary>
     internal static class LDtkSceneDrawerEntity
     {
-        public static ILDtkHandleDrawer DrawEntity(LDtkEntityDrawerData entity)
+        public static void DrawEntityDrawer(LDtkComponentEntity entityComponent)
         {
             Vector2 offset = Vector2.zero;
 
-            if (entity.Transform == null)
+            if (!entityComponent.Def)
             {
-                return null;
+                return;
             }
             
-            switch (entity.EntityMode)
+            switch (entityComponent.Def.RenderMode)
             {
                 case RenderMode.Cross:
                 case RenderMode.Ellipse:
                 case RenderMode.Rectangle:
                 case RenderMode.Tile:
-                    LDtkEntityDrawerShapes.Data shapeData = new LDtkEntityDrawerShapes.Data()
-                    {
-                        EntityMode = entity.EntityMode,
-                        FillOpacity = entity.FillOpacity,
-                        LineOpacity = entity.LineOpacity,
-                        Hollow = entity.Hollow,
-                        Pivot = entity.Pivot,
-                        Size = entity.Size
-                    };
-                    LDtkEntityDrawerShapes entityDrawer = new LDtkEntityDrawerShapes(entity.Transform, shapeData);
+                    LDtkEntityDrawerShapes entityDrawer = new LDtkEntityDrawerShapes(entityComponent);
                     entityDrawer.OnDrawHandles();
                     break;
                 default:
                     throw new ArgumentOutOfRangeException();
             }
 
-            offset = TryDrawTile(entity, offset);
-            TryDrawName(entity, offset);
+            offset = TryDrawTile(entityComponent, offset);
+            TryDrawName(entityComponent, offset);
 
-            return null;
+            return;
         }
 
-        private static void TryDrawName(LDtkEntityDrawerData entity, Vector2 offset)
+        private static void TryDrawName(LDtkComponentEntity entity, Vector2 offset)
         {
-            if (entity.ShowName && LDtkPrefs.ShowEntityIdentifier)
+            if (entity.Def.ShowName && LDtkPrefs.ShowEntityIdentifier)
             {
-                HandleUtil.DrawText(entity.Identifier, entity.Transform.position, entity.SmartColor, offset);
+                HandleUtil.DrawText(entity.Identifier, entity.transform.position, entity.SmartColor, offset);
             }
         }
 
-        private static Vector2 TryDrawTile(LDtkEntityDrawerData entity, Vector2 offset)
+        private static Vector2 TryDrawTile(LDtkComponentEntity entityComponent, Vector2 offset)
         {
-            if (!entity.DrawTile)
+            bool entityDrawsTile = entityComponent.Def.RenderMode == RenderMode.Tile;
+            bool fieldDrawsTile = entityComponent.Def.FieldDefs.Any(field => field.EditorDisplayMode == EditorDisplayMode.EntityTile);
+            
+            bool drawsTile = entityDrawsTile || fieldDrawsTile;
+            
+            if (!drawsTile)
             {
                 return offset;
             }
             
-            Texture2D tex = AssetDatabase.LoadAssetAtPath<Texture2D>(entity.TexPath);
-            if (tex == null)
+            Sprite tile = entityComponent.Tile;
+            if (!tile)
             {
                 return offset;
             }
             
-            LDtkEntityDrawerIcon iconDrawer = new LDtkEntityDrawerIcon(entity.Transform, tex, entity.TexRect);
+            LDtkEntityDrawerIcon iconDrawer = new LDtkEntityDrawerIcon(entityComponent.transform, tile.texture, tile.rect);
 
             iconDrawer.PrecalculateValues();
             offset += iconDrawer.OffsetToNextUI;