+++ Assets/LDtkUnity/Editor/CustomEditor/SceneDrawer/HandleDrawers/LDtkEntityDrawerIcon.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 namespace LDtkUnity.Editor
 {
-    internal sealed class LDtkEntityDrawerIcon : ILDtkHandleDrawer
+    internal sealed class LDtkEntityDrawerIcon
     {
         private readonly Transform _transform;
         private readonly Texture _tex;
@@ -93,6 +93,25 @@ namespace LDtkUnity.Editor
             GUI.color = prev;
 
             Handles.EndGUI();
+        }
+        
+        //todo incorporate this eventually to simplify the code
+        public static void DrawSprite(Rect position, Sprite sprite)
+        {
+            if (!sprite || !sprite.texture)
+            {
+                Debug.LogWarning("Sprite or its texture is null.");
+                return;
+            }
+
+            Rect texCoords = new Rect(
+                sprite.textureRect.x / sprite.texture.width,
+                sprite.textureRect.y / sprite.texture.height,
+                sprite.textureRect.width / sprite.texture.width,
+                sprite.textureRect.height / sprite.texture.height
+            );
+
+            GUI.DrawTextureWithTexCoords(position, sprite.texture, texCoords);
         }
     }
 }